+++ eslint.config.js
@@ -1,29 +1,41 @@
-import eslint from "@eslint/js";
 import prettier from "eslint-config-prettier";
+import js from "@eslint/js";
+import { includeIgnoreFile } from "@eslint/compat";
 import svelte from "eslint-plugin-svelte";
 import globals from "globals";
-import tseslint from "typescript-eslint";
+import { fileURLToPath } from "node:url";
+import ts from "typescript-eslint";
+import oxlint from "eslint-plugin-oxlint";
+const gitignorePath = fileURLToPath(new URL("./.gitignore", import.meta.url));
 
-export default tseslint.config(
-	eslint.configs.recommended,
-	...tseslint.configs.recommended,
-	...svelte.configs["flat/recommended"],
+export default ts.config(
+	includeIgnoreFile(gitignorePath),
+	js.configs.recommended,
+	...ts.configs.recommended,
+	...svelte.configs.recommended,
 	prettier,
-	...svelte.configs["flat/prettier"],
+	...svelte.configs.prettier,
 	{
 		languageOptions: {
-			globals: {
-				...globals.browser,
-				...globals.node,
-			},
+			globals: { ...globals.browser, ...globals.node },
 		},
+		rules: { "no-undef": "off" },
 	},
 	{
-		files: ["**/*.svelte"],
+		files: ["**/*.svelte", "**/*.svelte.ts", "**/*.svelte.js"],
+		ignores: ["eslint.config.js", "svelte.config.js"],
 		languageOptions: {
 			parserOptions: {
-				parser: tseslint.parser,
+				projectService: true,
+				extraFileExtensions: [".svelte"],
+				parser: ts.parser,
 			},
+		},
+		rules: {
+			"prefer-const": "off",
+			"svelte/no-navigation-without-resolve": "off",
+			"svelte/prefer-svelte-reactivity": "off",
+			"no-unused-private-class-members": "off",
 		},
 	},
 	{
@@ -37,7 +49,6 @@ export default tseslint.config(
 			],
 			"@typescript-eslint/no-unused-expressions": "off",
 			"@typescript-eslint/no-empty-object-type": "off",
-			"prefer-const": "off",
 		},
 	},
 	{
@@ -51,5 +62,6 @@ export default tseslint.config(
 			"packages/runed/dist/**/*",
 			"packages/runed/.svelte-kit/**/*",
 		],
-	}
+	},
+	...oxlint.configs["flat/recommended"]
 );