+++ packages/runed/src/lib/index.ts
@@ -1,1 +1,2 @@
 export * from "./utilities/index.js";
+export type { MaybeGetter, Getter, Setter } from "./internal/types.js";